context-tool/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.4.0
pyyaml>=6.0.1
sentence-transformers>=2.2.2
numpy>=1.24.0
pyperclip>=1.8.2
pynput>=1.7.6
pytest>=7.4.0

# Optional native UI dependency: Tkinter
# Tkinter is part of the Python standard library but may require an OS package
# to be installed on some platforms (not available via pip):
#   - Debian/Ubuntu: sudo apt install python3-tk
#   - Fedora: sudo dnf install python3-tkinter
#   - Arch: sudo pacman -S tk
#   - macOS: usually included with the official Python installer
# If you prefer a pip-installable GUI toolkit, uncomment/use PyQt6 instead:
# PyQt6>=6.6.0
